requirements.txt
# Core ML and RL libraries
torch>=1.12.0
torchvision>=0.13.0
stable-baselines3[extra]>=1.6.0  # Include extra packages for progress bars
gymnasium>=0.28.0
numpy>=1.21.0
opencv-python>=4.6.0

# Emulation and game interaction
pynput>=1.7.6
pillow>=9.0.0
mss>=6.1.0
pywin32>=305  # For Windows GUI interaction (win32gui)

# Data processing and visualization
pandas>=1.4.0
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.10.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0
rich>=12.0.0  # For progress bars
argparse
pathlib

# Optional: For advanced features
wandb>=0.13.0  # For experiment tracking
moviepy>=1.0.3  # For video recording
scikit-image>=0.19.0  # For image processing

# Development dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0 